+++ src/preprocess/preprocess_WASSA2021.py
@@ -20,7 +20,6 @@ import argparse
 # get label mapping for WASSA dataset
 label_mapping = fetch_label_mapping(wassa=True)
 
-
 def load_wassa_with_glove(tsv_path, max_length=128):
     """
     Load and preprocess the WASSA 2021 dataset from a TSV file.
@@ -117,8 +116,12 @@ if __name__ == "__main__":
     parser.add_argument(
         "--with_glove", action="store_true", help="Use GloVe embeddings"
     )
+    parser.add_argument(
+        "--force_preprocess", action="store_true", help="Force reprocessing even if files exist"
+    )
     args = parser.parse_args()
     with_glove = args.with_glove
+    force_preprocess = args.force_preprocess
 
     # Ensure the data folder exists
     os.makedirs("data", exist_ok=True)
@@ -126,24 +129,32 @@ if __name__ == "__main__":
     print("[Main] Loading and processing WASSA 2021 dataset...")
 
     if with_glove:
-        wassa_dataset, tokenizer = load_wassa_with_glove(WASSA_PATH, max_length=128)
-        load_glove_embeddings(tokenizer, WASSA_GLOVE_EMBEDDINGS_PATH)
+        if not force_preprocess and os.path.exists(WASSA_TRAIN_DS_PATH_WITH_GLOVE) and os.path.exists(WASSA_TEST_DS_PATH_WITH_GLOVE):
+            print("[Main] Dataset already preprocessed. Skipping...")
+        else:
+            print("[Main] Preprocessing Dataset.")
+            wassa_dataset, tokenizer = load_wassa_with_glove(WASSA_PATH, max_length=128)
+            load_glove_embeddings(tokenizer, WASSA_GLOVE_EMBEDDINGS_PATH)
 
-        print("[Main] Splitting dataset into train and test...")
-        train_ds, test_ds = split_dataset(wassa_dataset, split_ratio=0.8, glove=True)
+            print("[Main] Splitting dataset into train and test...")
+            train_ds, test_ds = split_dataset(wassa_dataset, split_ratio=0.8, glove=True)
 
-        print("[Main] Saving datasets to disk...")
-        torch.save(train_ds, WASSA_TRAIN_DS_PATH_WITH_GLOVE)
-        torch.save(test_ds, WASSA_TEST_DS_PATH_WITH_GLOVE)
+            print("[Main] Saving datasets to disk...")
+            torch.save(train_ds, WASSA_TRAIN_DS_PATH_WITH_GLOVE)
+            torch.save(test_ds, WASSA_TEST_DS_PATH_WITH_GLOVE)
 
     else:
-        wassa_dataset = load_wassa_without_glove(WASSA_PATH, max_length=128)
+        if not force_preprocess and os.path.exists(WASSA_TRAIN_DS_PATH_WITHOUT_GLOVE) and os.path.exists(WASSA_TEST_DS_PATH_WITHOUT_GLOVE):
+            print("[Main] Dataset already preprocessed. Skipping...")
+        else:
+            print("[Main] Preprocessing Dataset.")
+            wassa_dataset = load_wassa_without_glove(WASSA_PATH, max_length=128)
 
-        print("[Main] Splitting dataset into train and test...")
-        train_ds, test_ds = split_dataset(wassa_dataset, split_ratio=0.8, glove=False)
+            print("[Main] Splitting dataset into train and test...")
+            train_ds, test_ds = split_dataset(wassa_dataset, split_ratio=0.8, glove=False)
 
-        print("[Main] Saving datasets to disk...")
-        torch.save(train_ds, WASSA_TRAIN_DS_PATH_WITHOUT_GLOVE)
-        torch.save(test_ds, WASSA_TEST_DS_PATH_WITHOUT_GLOVE)
+            print("[Main] Saving datasets to disk...")
+            torch.save(train_ds, WASSA_TRAIN_DS_PATH_WITHOUT_GLOVE)
+            torch.save(test_ds, WASSA_TEST_DS_PATH_WITHOUT_GLOVE)
 
     print("[Main] Done.")
